pyproject.toml
[tool.poetry]
name = "великий проект"
version = "0.1.0"
description = ""
authors = ["czolowek <klutoi39@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
sqlalchemy = "^2.0.37"
psycopg2 = "^2.9.10"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.1"
flask-migrate = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
